requirements.txt
langchain
langchain_openai
langchain_core
python_dotenv
langserve
fastapi
uvicorn
google-generativeai
langchain-google-generativeai